refactor(SystemsDetails): WIP
