--- src/PresentationalComponents/SystemPolicyCards/index.js
@@ -1,2 +1,0 @@
-export { default } from './SystemPolicyCards';
-export * from './SystemPolicyCards';
