+++ src/PresentationalComponents/index.js
@@ -57,3 +57,4 @@ export { default as InUseProfileLabel } from './InUseProfileLabel/InUseProfileLa
 export { default as WithPermission } from './WithPermission/WithPermission';
 export { default as LinkWithPermission } from './LinkWithPermission/LinkWithPermission';
 export { default as LinkButton } from './LinkButton/LinkButton';
+export { default as ComplianceEmptyState } from './ComplianceEmptyState/ComplianceEmptyState';
