+++ src/SmartComponents/SystemDetails/components/EmptyState/EmptyState.js
