+++ src/SmartComponents/SystemDetails/components/EmptyState/NoPoliciesState.js
