+++ src/SmartComponents/SystemDetails/components/EmptyState/NoPoliciesState.test.js
