+++ src/SmartComponents/SystemDetails/components/EmptyState/NoReportsState.js
