+++ src/SmartComponents/SystemDetails/components/EmptyState/NoReportsState.test.js
