+++ src/SmartComponents/SystemDetails/components/EmptyState/index.js
@@ -1,0 +1,1 @@
+export { default } from './EmptyState';
