+++ src/SmartComponents/SystemDetails/components/SystemPolicyCard/SystemPolicyCard.test.js
