+++ src/SmartComponents/SystemDetails/components/SystemPolicyCard/UnsupportedSSGVersion.js
