+++ src/SmartComponents/SystemDetails/components/SystemPolicyCard/index.js
