+++ src/SmartComponents/index.js
@@ -1,5 +1,4 @@
 export { default as ComplianceSystems } from './ComplianceSystems/ComplianceSystems';
-export { default as InventoryDetails } from './InventoryDetails/InventoryDetails';
 export { default as CreatePolicy } from './CreatePolicy/CreatePolicy';
 export { default as DeleteReport } from './DeleteReport/DeleteReport';
 export { default as EditPolicy } from './EditPolicy/EditPolicy';
