+++ src/Utilities/hooks/useDocumentTitle.js
@@ -3,7 +3,8 @@ import { useLocation } from 'react-router-dom';
 
 import { DEFAULT_TITLE, DEFAULT_TITLE_SUFFIX } from '@/constants';
 
-const defaultTitleSuffixed = (title) => `${title}${DEFAULT_TITLE_SUFFIX}`;
+export const defaultTitleSuffixed = (title = DEFAULT_TITLE) =>
+  `${title}${DEFAULT_TITLE_SUFFIX}`;
 
 export const useTitleEntity = (route, entityTitle) => {
   const location = useLocation();
